+++ src/main/java/org/truffleruby/cext/CExtNodes.java
@@ -1313,7 +1313,7 @@ public abstract class CExtNodes {
         Object rbSysErrFail(int errno, Object string,
                 @Cached ErrnoErrorNode errnoErrorNode) {
             final Backtrace backtrace = getContext().getCallStack().getBacktrace(this);
-            throw new RaiseException(getContext(), errnoErrorNode.execute(null, errno, string, backtrace));
+            throw new RaiseException(getContext(), errnoErrorNode.execute(null, errno, string, null, backtrace));
         }
 
     }